requirements.txt
google-api-python-client==2.90.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
helium==3.2.2
python-dateutil==2.8.2
